+++ packages/s2-core/src/utils/export/copy.ts
@@ -139,7 +139,7 @@ const format = (
 export const convertString = (v: string) => {
   if (/\n/.test(v)) {
     // 单元格内换行 替换双引号 防止内容存在双引号 导致内容换行出错
-    return '"' + v.replace(/\r\n?/g, '\n').replace(/"/g, "'") + '"';
+    return '"' + v.replace(/\r\n?/g, '\n') + '"';
   }
   return v;
 };
@@ -175,31 +175,6 @@ export type CopyableItem = {
 
 export type Copyable = CopyableItem | CopyableItem[];
 
-function pickDataFromCopyable(
-  copyable: Copyable,
-  type: CopyMIMEType[],
-): string[];
-function pickDataFromCopyable(copyable: Copyable, type: CopyMIMEType): string;
-function pickDataFromCopyable(
-  copyable: Copyable,
-  type: CopyMIMEType | CopyMIMEType[],
-): string | string[];
-function pickDataFromCopyable(
-  copyable: Copyable,
-  type: CopyMIMEType[] | CopyMIMEType = CopyMIMEType.PLAIN,
-): string[] | string {
-  if (Array.isArray(type)) {
-    return ([].concat(copyable) as CopyableItem[])
-      .filter((item) => type.includes(item.type))
-      .map((item) => item.content);
-  }
-  return (
-    ([].concat(copyable) as CopyableItem[])
-      .filter((item) => item?.type === type)
-      .map((item) => item.content)[0] || ''
-  );
-}
-
 // 把 string[][] 矩阵转换成 CopyableItem
 const matrixPlainTextTransformer: MatrixTransformer = (dataMatrix) => {
   return {
@@ -230,6 +205,49 @@ const matrixHtmlTransformer: MatrixTransformer = (dataMatrix) => {
     )}</tbody></table>`,
   };
 };
+
+const transformers: {
+  [key in CopyMIMEType]: MatrixTransformer;
+} = {
+  [CopyMIMEType.PLAIN]: matrixPlainTextTransformer,
+  [CopyMIMEType.HTML]: matrixHtmlTransformer,
+};
+
+export function registerTransformer(
+  type: CopyMIMEType,
+  transformer: MatrixTransformer,
+) {
+  transformers[type] = transformer;
+}
+
+function getTransformer(type: CopyMIMEType) {
+  return transformers[type];
+}
+
+function pickDataFromCopyable(
+  copyable: Copyable,
+  type: CopyMIMEType[],
+): string[];
+function pickDataFromCopyable(copyable: Copyable, type: CopyMIMEType): string;
+function pickDataFromCopyable(
+  copyable: Copyable,
+  type: CopyMIMEType | CopyMIMEType[],
+): string | string[];
+function pickDataFromCopyable(
+  copyable: Copyable,
+  type: CopyMIMEType[] | CopyMIMEType = CopyMIMEType.PLAIN,
+): string[] | string {
+  if (Array.isArray(type)) {
+    return ([].concat(copyable) as CopyableItem[])
+      .filter((item) => type.includes(item.type))
+      .map((item) => item.content);
+  }
+  return (
+    ([].concat(copyable) as CopyableItem[])
+      .filter((item) => item?.type === type)
+      .map((item) => item.content)[0] || ''
+  );
+}
 
 // 生成矩阵：https://gw.alipayobjects.com/zos/antfincdn/bxBVt0nXx/a182c1d4-81bf-469f-b868-8b2e29acfc5f.png
 const assembleMatrix = (
@@ -269,7 +287,10 @@ const assembleMatrix = (
     });
   }) as string[][];
 
-  return [matrixPlainTextTransformer(matrix), matrixHtmlTransformer(matrix)];
+  return [
+    getTransformer(CopyMIMEType.PLAIN)(matrix),
+    getTransformer(CopyMIMEType.HTML)(matrix),
+  ];
 };
 
 export const processCopyData = (
@@ -278,10 +299,18 @@ export const processCopyData = (
   spreadsheet: SpreadSheet,
 ): Copyable => {
   const matrix = cells.map((cols) =>
-    cols.map((item) => convertString(format(item, displayData, spreadsheet))),
+    cols.map((item) => {
+      if (!item) {
+        return '';
+      }
+      return convertString(format(item, displayData, spreadsheet));
+    }),
   );
 
-  return [matrixPlainTextTransformer(matrix), matrixHtmlTransformer(matrix)];
+  return [
+    getTransformer(CopyMIMEType.PLAIN)(matrix),
+    getTransformer(CopyMIMEType.HTML)(matrix),
+  ];
 };
 
 /**
@@ -342,15 +371,15 @@ const processTableColSelected = (
           meta: node,
         }));
 
-  const dataMatrix = displayData.map((row) => {
+  const matrix = displayData.map((row) => {
     return selectedFields.map(({ field, formatter, meta }) =>
       convertString(formatter(row[field], undefined, meta)),
     );
   });
 
   return [
-    matrixPlainTextTransformer(dataMatrix),
-    matrixHtmlTransformer(dataMatrix),
+    getTransformer(CopyMIMEType.PLAIN)(matrix),
+    getTransformer(CopyMIMEType.HTML)(matrix),
   ];
 };
 
@@ -386,10 +415,11 @@ const getPivotWithoutHeaderCopyData = (
   leafRows: Node[],
   leafCols: Node[],
 ): Copyable => {
-  const dataMatrix = getDataMatrix(leafRows, leafCols, spreadsheet);
+  const matrix = getDataMatrix(leafRows, leafCols, spreadsheet);
+
   return [
-    matrixPlainTextTransformer(dataMatrix),
-    matrixHtmlTransformer(dataMatrix),
+    getTransformer(CopyMIMEType.PLAIN)(matrix),
+    getTransformer(CopyMIMEType.HTML)(matrix),
   ];
 };
 
@@ -466,7 +496,11 @@ const processTableRowSelected = (
           ),
         );
     });
-  return [matrixPlainTextTransformer(matrix), matrixHtmlTransformer(matrix)];
+
+  return [
+    getTransformer(CopyMIMEType.PLAIN)(matrix),
+    getTransformer(CopyMIMEType.HTML)(matrix),
+  ];
 };
 
 const processPivotRowSelected = (
@@ -666,15 +700,15 @@ function getBrushHeaderCopyable(
 
   // 拼接选中行列头的内容矩阵
   const isCol = interactedCells[0].cellType === CellTypes.COL_CELL;
-  let cellMatrix = getCellMatrix(lastLevelCells, maxLevel, allLevels);
+  let matrix = getCellMatrix(lastLevelCells, maxLevel, allLevels);
 
   // 如果是列头，需要转置
   if (isCol) {
-    cellMatrix = zip(...cellMatrix);
+    matrix = zip(...matrix);
   }
   return [
-    matrixPlainTextTransformer(cellMatrix),
-    matrixHtmlTransformer(cellMatrix),
+    getTransformer(CopyMIMEType.PLAIN)(matrix),
+    getTransformer(CopyMIMEType.HTML)(matrix),
   ];
 }
 
@@ -698,7 +732,7 @@ export const getDataByRowData = (
   } = spreadsheet;
   const defaultDataValue = getEmptyPlaceholder(spreadsheet, placeholder);
   const column = spreadsheet.getColumnLeafNodes();
-  let datas: string[][] = [];
+  let matrix: string[][] = [];
 
   if (spreadsheet.isTableMode()) {
     const columnWithoutSeriesNumber = filter(
@@ -706,7 +740,7 @@ export const getDataByRowData = (
       (node) => node.field !== SERIES_NUMBER_FIELD,
     );
     // 按列头顺序复制
-    datas = map(rowData, (rowDataItem) => {
+    matrix = map(rowData, (rowDataItem) => {
       return map(
         columnWithoutSeriesNumber,
         (node) => rowDataItem?.[node.field] ?? defaultDataValue,
@@ -720,7 +754,7 @@ export const getDataByRowData = (
     const rowDataFlattenWithoutTotal = rowDataFlatten.filter((data) =>
       [...rows, ...columns].every((field) => !isNil(data[field as string])),
     );
-    datas = reduce(
+    matrix = reduce(
       rowDataFlattenWithoutTotal,
       (ret, data) => {
         return [...ret, tilePivotData(data, columnOrdered, defaultDataValue)];
@@ -728,7 +762,10 @@ export const getDataByRowData = (
       [],
     );
   }
-  return matrixPlainTextTransformer(datas);
+  return [
+    getTransformer(CopyMIMEType.PLAIN)(matrix),
+    getTransformer(CopyMIMEType.HTML)(matrix),
+  ];
 };
 
 function getDataCellCopyable(
@@ -739,7 +776,6 @@ function getDataCellCopyable(
 
   const selectedCols = cells.filter(({ type }) => type === CellTypes.COL_CELL);
   const selectedRows = cells.filter(({ type }) => type === CellTypes.ROW_CELL);
-
   const displayData = spreadsheet.dataSet.getDisplayDataSet();
 
   if (