+++ packages/s2-react/playground/index.tsx
@@ -259,12 +259,12 @@ function MainLayout() {
   };
 
   const logHandler =
-    (name: string, callback?: () => void) =>
-    (...args: unknown[]) => {
+    (name: string, callback?: (...args: any[]) => void) =>
+    (...args: any[]) => {
       if (s2Ref.current?.options?.debug) {
         console.log(name, ...args);
       }
-      callback?.();
+      callback?.(...args);
     };
 
   const onColCellClick = (cellInfo: TargetCellInfo) => {
@@ -603,6 +603,20 @@ function MainLayout() {
                 </Button>
                 <Button
                   size="small"
+                  onClick={() => {
+                    clearInterval(scrollTimer.current);
+                    s2Ref.current.updateScrollOffset({
+                      rowHeaderOffsetX: {
+                        value: 100,
+                        animate: true,
+                      },
+                    });
+                  }}
+                >
+                  滚动行头
+                </Button>
+                <Button
+                  size="small"
                   danger
                   onClick={() => {
                     if (
@@ -823,7 +837,9 @@ function MainLayout() {
                   onChange={(checked) => {
                     updateOptions({
                       interaction: {
-                        linkFields: checked ? ['province', 'city'] : [],
+                        linkFields: checked
+                          ? ['province', 'city', 'number']
+                          : [],
                       },
                     });
                   }}
@@ -1521,7 +1537,10 @@ function MainLayout() {
               })}
               onColCellClick={onColCellClick}
               onRowCellClick={logHandler('onRowCellClick')}
-              onCornerCellClick={(cellInfo) => {
+              onCornerCellClick={logHandler('onCornerCellClick', (cellInfo) => {
+                if (!showCustomTooltip) {
+                  return;
+                }
                 s2Ref.current.showTooltip({
                   position: {
                     x: cellInfo.event.clientX,
@@ -1529,7 +1548,7 @@ function MainLayout() {
                   },
                   content: 'click',
                 });
-              }}
+              })}
               onDataCellClick={logHandler('onDataCellClick')}
               onLayoutResize={logHandler('onLayoutResize')}
               onCopied={logHandler('onCopied')}