+++ s2-site/.dumi/pages/playground.zh.tsx
@@ -6,11 +6,11 @@ import { Footer } from '@tant/dumi-theme-antv/dist/slots/Footer';
 const Playground: React.FC = () => {
   return (
     <>
-    <Header/>
-    <Page />
-    <Footer />
+      <Header />
+      <Page />
+      <Footer />
     </>
-  )
+  );
 };
 
 export default Playground;